+++ app/src/androidTest/java/com/battlelancer/seriesguide/provider/DefaultValuesTest.java
@@ -128,6 +128,8 @@ public class DefaultValuesTest {
         assertThat(show.getCustomReleaseTime()).isNull();
         assertThat(show.getCustomReleaseDayOffset()).isNull();
         assertThat(show.getCustomReleaseTimeZone()).isNull();
+        assertThat(show.getUserNote()).isNull();
+        assertThat(show.getUserNoteTraktId()).isNull();
     }
 
     @Test