+++ app/src/main/java/com/battlelancer/seriesguide/shows/ShowsActivityImpl.kt
@@ -5,7 +5,6 @@ package com.battlelancer.seriesguide.shows
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.KeyEvent
 import android.view.View
 import androidx.activity.viewModels
 import androidx.lifecycle.ViewModelProvider
@@ -188,9 +187,9 @@ open class ShowsActivityImpl : BaseTopActivity() {
     }
 
     private fun setupViews() {
-        // setup floating action button for adding shows
-        val buttonAddShow = findViewById<FloatingActionButton>(R.id.buttonShowsAdd)
-        buttonAddShow.setOnClickListener {
+        // setup floating action button
+        val floatingActionButton = findViewById<FloatingActionButton>(R.id.buttonShowsFloating)
+        floatingActionButton.setOnClickListener {
             startActivity(ShowsDiscoverPagingActivity.intentSearch(this))
         }
 
@@ -205,7 +204,7 @@ open class ShowsActivityImpl : BaseTopActivity() {
         tabs.setOnPageChangeListener(
             ShowsPageChangeListener(
                 findViewById(R.id.sgAppBarLayout),
-                buttonAddShow,
+                floatingActionButton,
                 viewModel
             )
         )
@@ -326,11 +325,6 @@ open class ShowsActivityImpl : BaseTopActivity() {
         ShowsSettings.saveLastShowsTabPosition(this, viewPager.currentItem)
     }
 
-    override fun onKeyLongPress(keyCode: Int, event: KeyEvent): Boolean {
-        // prevent navigating to top activity as this is the top activity
-        return keyCode == KeyEvent.KEYCODE_BACK
-    }
-
     override val snackbarParentView: View
         get() = findViewById(R.id.coordinatorLayoutShows)
 
@@ -363,7 +357,6 @@ open class ShowsActivityImpl : BaseTopActivity() {
             }
             appBarLayout.liftOnScrollTargetViewId = liftOnScrollTarget
 
-            // only display add show button on Shows tab
             if (position == Tab.DISCOVER.index) {
                 floatingActionButton.show()
             } else {