+++ app/src/main/java/com/battlelancer/seriesguide/traktapi/TraktTools2.kt
@@ -1,11 +1,10 @@
-// Copyright 2023 Uwe Trottmann
 // SPDX-License-Identifier: Apache-2.0
+// Copyright 2021-2024 Uwe Trottmann
 
 package com.battlelancer.seriesguide.traktapi
 
 import android.content.Context
 import com.battlelancer.seriesguide.SgApp
-import com.battlelancer.seriesguide.shows.tools.AddUpdateShowTools.ShowResult
 import com.battlelancer.seriesguide.util.Errors
 import com.battlelancer.seriesguide.util.isRetryError
 import com.github.michaelbull.result.Err
@@ -14,20 +13,126 @@ import com.github.michaelbull.result.Result
 import com.github.michaelbull.result.andThen
 import com.github.michaelbull.result.mapError
 import com.github.michaelbull.result.runCatching
+import com.uwetrottmann.trakt5.TraktV2
+import com.uwetrottmann.trakt5.entities.AddNoteRequest
 import com.uwetrottmann.trakt5.entities.BaseShow
-import com.uwetrottmann.trakt5.entities.LastActivities
 import com.uwetrottmann.trakt5.entities.LastActivity
 import com.uwetrottmann.trakt5.entities.LastActivityMore
+import com.uwetrottmann.trakt5.entities.LastActivityUpdated
+import com.uwetrottmann.trakt5.entities.Note
 import com.uwetrottmann.trakt5.entities.Ratings
 import com.uwetrottmann.trakt5.entities.Show
+import com.uwetrottmann.trakt5.entities.ShowIds
 import com.uwetrottmann.trakt5.enums.Extended
 import com.uwetrottmann.trakt5.enums.IdType
 import com.uwetrottmann.trakt5.enums.Type
+import com.uwetrottmann.trakt5.services.Notes
+import retrofit2.Call
 import retrofit2.Response
+import retrofit2.awaitResponse
 
 object TraktTools2 {
 
-    data class SearchResult(val result: ShowResult, val show: Show?)
+    sealed interface TraktResponse<T> {
+        data class Success<T>(
+            /**
+             * If T is [Void] this is always `null`.
+             */
+            val data: T?
+        ) : TraktResponse<T>
+    }
+
+    sealed interface TraktNonNullResponse<T> {
+        data class Success<T>(val data: T) : TraktNonNullResponse<T>
+    }
+
+    sealed interface TraktErrorResponse {
+        class IsNotVip<T> : TraktResponse<T>, TraktNonNullResponse<T>
+        class IsUnauthorized<T> : TraktResponse<T>, TraktNonNullResponse<T>
+        class Other<T> : TraktResponse<T>, TraktNonNullResponse<T>
+    }
+
+    /**
+     * Adds or updates the note for the given show.
+     */
+    suspend fun saveNoteForShow(
+        traktNotes: Notes,
+        showTmdbId: Int,
+        noteText: String
+    ): TraktNonNullResponse<Note> {
+        // Note: calling the add endpoint for an existing note will update it
+        return awaitTraktCallNonNull(
+            traktNotes.addNote(
+                AddNoteRequest(
+                    Show().apply {
+                        ids = ShowIds.tmdb(showTmdbId)
+                    },
+                    noteText
+                )
+            ), "update note"
+        )
+    }
+
+    suspend fun deleteNote(
+        trakt: SgTrakt,
+        noteId: Long
+    ): TraktResponse<Void> {
+        return awaitTraktCall(trakt.notes().deleteNote(noteId), "delete note")
+    }
+
+    private suspend fun <T> awaitTraktCall(
+        call: Call<T>,
+        action: String,
+        logErrorOnNullBody: Boolean = false
+    ): TraktResponse<T> {
+        val response = try {
+            call.awaitResponse()
+        } catch (e: Exception) {
+            Errors.logAndReport(action, e)
+            return TraktErrorResponse.Other()
+        }
+
+        if (!response.isSuccessful) {
+            return when {
+                TraktV2.isNotVip(response) -> TraktErrorResponse.IsNotVip()
+                TraktV2.isUnauthorized(response) -> TraktErrorResponse.IsUnauthorized()
+                else -> {
+                    Errors.logAndReport(action, response)
+                    TraktErrorResponse.Other()
+                }
+            }
+        }
+
+        val body = response.body()
+
+        if (logErrorOnNullBody && body == null) {
+            Errors.logAndReport(action, response, "body is null")
+        }
+
+        return TraktResponse.Success(body)
+    }
+
+    /**
+     * Like [awaitTraktCall], but ensures the response data is not null.
+     */
+    private suspend fun <T> awaitTraktCallNonNull(
+        call: Call<T>,
+        action: String
+    ): TraktNonNullResponse<T> {
+        return when (val response = awaitTraktCall(call, action, logErrorOnNullBody = true)) {
+            is TraktErrorResponse.Other -> response
+            is TraktErrorResponse.IsNotVip -> response
+            is TraktErrorResponse.IsUnauthorized -> response
+            is TraktResponse.Success -> {
+                val data = response.data
+                if (data == null) {
+                    TraktErrorResponse.Other()
+                } else {
+                    TraktNonNullResponse.Success(data)
+                }
+            }
+        }
+    }
 
     /**
      * Look up a show by its TMDB ID, may return `null` if not found.
@@ -140,6 +245,7 @@ object TraktTools2 {
         val episodes: LastActivityMore,
         val shows: LastActivity,
         val movies: LastActivityMore,
+        val notes: LastActivityUpdated,
     )
 
     fun getLastActivity(context: Context): Result<LastActivities, TraktError> {
@@ -157,12 +263,14 @@ object TraktTools2 {
                 val episodes = lastActivities?.episodes
                 val shows = lastActivities?.shows
                 val movies = lastActivities?.movies
-                if (episodes != null && shows != null && movies != null) {
+                val notes = lastActivities?.notes
+                if (episodes != null && shows != null && movies != null && notes != null) {
                     return@andThen Ok(
                         LastActivities(
                             episodes = episodes,
                             shows = shows,
-                            movies = movies
+                            movies = movies,
+                            notes = notes
                         )
                     )
                 } else {