+++ app/src/main/java/com/battlelancer/seriesguide/util/WebTools.kt
@@ -1,5 +1,5 @@
-// Copyright 2023 Uwe Trottmann
 // SPDX-License-Identifier: Apache-2.0
+// Copyright 2023-2024 Uwe Trottmann
 
 package com.battlelancer.seriesguide.util
 
@@ -22,7 +22,6 @@ object WebTools {
      *
      * See also [openInApp].
      */
-    @JvmStatic
     fun openInCustomTab(context: Context, url: String): Boolean {
         val darkParams = CustomTabColorSchemeParams.Builder()
             .setToolbarColor(